+++ core-lib/es6/src/main/java/def/dom/Touch.java
@@ -1,15 +1,13 @@
 package def.dom;
 
-import def.js.Object;
-
 public class Touch extends def.js.Object {
-    public double clientX;
-    public double clientY;
-    public double identifier;
-    public double pageX;
-    public double pageY;
-    public double screenX;
-    public double screenY;
+    public int clientX;
+    public int clientY;
+    public int identifier;
+    public int pageX;
+    public int pageY;
+    public int screenX;
+    public int screenY;
     public EventTarget target;
     public static Touch prototype;
     public Touch(){}
